% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icg}
\alias{icg}
\title{Inversive Congruential Generator}
\usage{
icg(q, a, c, seed, n)
}
\arguments{
\item{q}{TODO}

\item{a}{TODO}

\item{c}{TODO}

\item{seed}{TODO}

\item{n}{}
}
\description{
Inversive congruential generators are a type of non-linear congregational pseudorandom number generator which use the modular multiplicative inverse (if it exists) to generate the next number in a sequence. \br\br The standard formula for an inversive congruential generator, modulo some prime \eqn{q} is: \br \eqn{x_0 = \text{seed}, \quad x_{i+1} = \left\{ \begin{array}{ll} (a x_i^{-1} + c) \mod q & \text{if } x_i \neq 0, \\ c & \text{if } x_i = 0. \end{array} \right}
}
\details{
For more information, please see the \href{https://en.wikipedia.org/wiki/Inversive_congruential_generator}{Wikipedia Page}.
}
