% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcg}
\alias{lcg}
\title{Linear Congruential Generator}
\usage{
lcg(seed, n)
}
\arguments{
\item{seed}{initial starting value}

\item{n}{the number of random numbers you want to create.}
}
\description{
[TODO: WRITE DESCRIPTION]
}
\details{
For more reading check out Wikipedia:https://en.wikipedia.org/wiki/Linear_congruential_generator
}
\examples{
random_numbers <- lcg(1234, 1000)
# Plot numbers to see that they are random
plot(random_numbers)
}
