% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcg_parkmiller}
\alias{lcg_parkmiller}
\title{Park Miller Generator}
\usage{
lcg_parkmiller(seed, n)
}
\arguments{
\item{seed}{initial starting value}

\item{n}{the number of random numbers you want to create.}
}
\description{
TODO: Documentation
}
\details{
For more information see the \href{https://en.wikipedia.org/wiki/Lehmer_random_number_generator}{Wikipedia page}.
}
\examples{
random_numbers <- lcg_parkmiller(1234, 1000)
# Plot numbers to see that they are random
plot(random_numbers)
}
