% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcg_tos}
\alias{lcg_tos}
\title{TempleOS Linear Congruential Generator (LCG)}
\usage{
lcg_tos(n = 1L)
}
\arguments{
\item{n}{Number of random numbers to generate (default 1).}
}
\value{
A NumericVector of length `n`. Each value is the 64-bit state coerced to double.
}
\description{
Generates `n` pseudo-random values using the TempleOS constants, with an initial seed taken from the CPU timestamp counter. This code is inspired from the Linear Congruential Generator Specified in the TempleOS source code. To see the source code in TempleOS, see \href{https://tinkeros.github.io/WbTempleOS/Kernel/KMathB.html#l77}{here}.
}
\details{
\code{lcg_tos} differs from \code{lcg} as the parameters are preconfigured and the random seed comes from the CPU timestamp. It is possible to configure \code{lcg_tos} with \code{lcg}, however this is ready to use preconfigured function that only requires the user to specify the number of random numbers they wish to generate.
}
\examples{
random_numbers <- lcg_tos(10000)
# Plot numbers to see that they are random
plot(random_numbers)
}
\seealso{
\code{\link{lcg}}
}
