% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lfg}
\alias{lfg}
\title{Lagged Fibonacci Generator (LFG)}
\usage{
lfg(n, j = 24L, k = 55L, bitsize = 32L, operation = "+")
}
\arguments{
\item{n}{Number of random numbers to generate.}

\item{j}{Lag parameter j (0 < j < k).}

\item{k}{Lag parameter k (j < k).}

\item{bitsize}{Bit size for modulo (m = 2^bitsize).}

\item{operation}{Character representing the operation: '+', '-', '*', '^' (XOR).}
}
\value{
NumericVector of generated values.
}
\description{
A Lagged Fibbonacci Generator (LFG or sometimes LFib) is an example of a pseudorandom number generator. This class of pseudorandom number generator is aimed at being an improvment of the 'standard' linear congruential generator.  These are based on the generalization of the Fibbonacci sequence.\cr\cr
The Fibonacci sequence may be described by the recurrence relation:\cr\cr
\eqn{S_n = S_{n-1}+S_{n-2}}\cr\cr
Hence, the new term is the sum of the last two terms in the sequence This can be generalized by the sequence:\cr\cr
\eqn{S_n  \equiv S_{n-j} \ast S_{n-k} \space (\text{mod }m), 0 < j <k}\cr\cr
In which case, the new term is some combination of any two previous terms. m is usually a power of 2 (m = 2M), often 232 or 264. The  \eqn{\ast} operator denotes a general binary operation. This may be either addition, subtraction, multiplication, or the bitwise exclusive-or operator (XOR). In this package, the \eqn{\ast} operator denotes addition.The theory of this type of generator is rather complex, and it may not be sufficient simply to choose random values for j and k. These generators also tend to be very sensitive to initialisation.
}
\details{
For more information, see the \href{https://en.wikipedia.org/wiki/Lagged_Fibonacci_generator}{Wikipedia page}.
}
\examples{
plot(
lfg(10000,j=7083 ,k=19937, bitsize=32, operation = '^')
  )
}
