% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lfsr_xorshift}
\alias{lfsr_xorshift}
\title{Xorshift Linear Feedback Shift Register (LFSR)}
\usage{
lfsr_xorshift(seed, n, bitsize, taps)
}
\description{
For more information, check out the \href{https://en.wikipedia.org/wiki/Linear-feedback_shift_register}{Wikipedia Page}.
}
