% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{winchmann_hill}
\alias{winchmann_hill}
\title{Winchman-Hill}
\usage{
winchmann_hill(seed1, seed2, seed3, n)
}
\arguments{
\item{seed1}{Initial seed value.}

\item{seed2}{Initial seed value.}

\item{seed3}{Initial seed value.}

\item{n}{Number of random numbers to generate.}
}
\description{
The Wichmann-Hill random number generator indeed produces pseudorandom numbers that are typically in the range \eqn{\left[0, 1\right)}. It generates floating-point numbers rather than discrete integers.
}
\details{
For more information see the \href{https://en.wikipedia.org/wiki/Wichmann–Hill}{Wikipedia Page}
}
\examples{
random_numbers <- winchmann_hill(1234, 4321,5678,1000)
# Plot numbers to see that they are random
plot(random_numbers)
}
