% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmwc}
\alias{cmwc}
\title{Complementary Multiply with Carry Generator}
\usage{
cmwc(seed, n, r, bitsize = 32L, a = 7L, c = 4L)
}
\description{
Wikipedia: https://en.wikipedia.org/wiki/Multiply-with-carry_pseudorandom_number_generator#Complementary-multiply-with-carry_generators
}
