% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{icg}
\alias{icg}
\title{Inversive Congruential Generator}
\usage{
icg(q, a, c, seed, n)
}
\arguments{
\item{q}{TODO}

\item{a}{TODO}

\item{c}{TODO}

\item{seed}{TODO}

\item{n}{}
}
\description{
NEEDS DOCUMENTATION
}
