% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lfg}
\alias{lfg}
\title{Lagged Fibonacci Generator (LFG)}
\usage{
lfg(seed, n, j = 24L, k = 55L, bitsize = 32L)
}
\arguments{
\item{seed}{TODO}

\item{j}{TODO}

\item{k}{TODO}

\item{int}{n TODO}
}
\description{
[DESCRIPTION: TODO]
}
\examples{
random_numbers <- lfg(13124, 10000)
# Plot numbers to see that they are random
plot(random_numbers)
}
