% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lfsr_fib}
\alias{lfsr_fib}
\title{Linear Feedback Shift Register}
\usage{
lfsr_fib(seed, n)
}
\description{
For more information, check out the \href{https://en.wikipedia.org/wiki/Linear-feedback_shift_register}{Wikipedia Page}.
}
