% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{middlesquare}
\alias{middlesquare}
\title{Middles-Square Method}
\usage{
middlesquare(seed, n)
}
\arguments{
\item{seed}{initial starting value}

\item{n}{number of random numbers desired}
}
\description{
[DESCRIPTION: TODO]
This code is based on the python code shared on Wikipedia ([link](https://en.wikipedia.org/wiki/Middle-square_method))
}
\examples{
random_numbers <- middlesquare(1234, 1000)
# Plot numbers to see that they are random
plot(random_numbers)
}
