% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_randtests.R
\name{run_randtests}
\alias{run_randtests}
\title{Comprehensive Randomness Battery}
\usage{
run_randtests(x, alter = "two.sided")
}
\arguments{
\item{x}{A numeric vector}

\item{alter}{Alternative hypothesis: "two.sided" (default), "less", or "greater"}
}
\value{
A data.frame containing detailed diagnostic results
}
\description{
Comprehensive Randomness Battery
}
